Version 1.4 rework
