+++ src/HeatMap/HeatMapKeyBindings.cs
@@ -1,0 +1,16 @@
+﻿using RimWorld;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Verse;
+
+namespace HeatMap
+{
+    [DefOf]
+    public static class HeatMapKeyBingings
+    {
+        public static KeyBindingDef ToggleHeatMap;
+    }
+}
