--- src/HeatMap/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Lib.Harmony" version="2.1.0" targetFramework="net472" />
-  <package id="UnlimitedHugs.Rimworld.HugsLib" version="9.0.0" targetFramework="net472" />
-</packages>
